+++ src/commands/installDeps.ts
@@ -1,39 +1,36 @@
-import { exec } from "child_process";
-import { promisify } from "util";
-
-const execPromise = promisify(exec);
+import { pkgManagerInstall, pkgManagerAdd } from "./runner";
 
 async function install(
   projectName: string,
   dependencies: string,
+  pkgManager: string,
   verbose?: boolean,
 ) {
   console.log(`installing: ${dependencies}`);
-
-  try {
-    const { stdout, stderr } = await execPromise(
-      `pnpm install ${dependencies}`,
-      {
-        cwd: projectName,
-      },
-    );
-
-    console.log(`Successfully installed: ${dependencies}`);
-
-    if (verbose == true) {
-      console.log(`stdout: ${stdout}`);
-      console.log(`stderr: ${stderr}`);
-    }
-  } catch (error) {
-    console.error(`error installing ${dependencies}` + error);
+  if (pkgManager == "pnpm" || pkgManager == "npm") {
+    await pkgManagerInstall(projectName, dependencies, pkgManager, verbose);
+  }
+  if (pkgManager == "yarn") {
+    await pkgManagerAdd(projectName, dependencies, pkgManager, verbose);
   }
 }
 
-export async function install_deps(projectName: string, verbose?: boolean) {
-  await install(projectName, "express cors body-parser dotenv morgan", verbose);
-  await install(
-    projectName,
-    "swagger-autogen swagger-ui-express prettier -D",
-    verbose,
-  );
+export async function install_deps(
+  projectName: string,
+  programmingLanguage: string,
+  pkgManager: string,
+  verbose?: boolean,
+) {
+  const deps = "express cors body-parser dotenv morgan";
+  const devDeps = "swagger-autogen swagger-ui-express prettier -D";
+  await install(projectName, deps, pkgManager, verbose);
+  await install(projectName, devDeps, pkgManager, verbose);
+  if (programmingLanguage == "typescript") {
+    await install(
+      projectName,
+      "typescript tsx nodemon @types/node @types/express @types/cors @types/body-parser @types/morgan @types/pg -D",
+      pkgManager,
+      verbose,
+    );
+  }
 }