+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # CHANGELOG
 
-## [0.2.0]
+## [0.3.0] 2024-05-15
+### added
+- choose between typescript and javascript
+- choose between your favorite package manager
+
+### fixed
+- mapi installs @types for libs now
+
+## [0.2.0] 2024-05-12
 
 ### added
 
@@ -18,7 +26,7 @@
 
 - fixed "command not found"
 
-## [1.0.0] - 2024-04-01
+## [1.0.0] - 2024-05-10
 
 ### Added